changed to structures inside arrays / fixed procs
I changed all previous procedures to make them able to use structures
with arrays, this helps cut down a majority of the code and is also a
little easier to follow and to make new procedures.  I fixed the adding
card to player procedure, there was some size mismatching that caused it
to not work and made it look like other things weren't working.  I will
start tomorrow working on the game play code and see how far i can get.
if anyone wants to try and do the hand value proc that'd be great, all
we really need to do is make the game code and implement a gui,
everything else seems to be working good.. i checked all the procedures
after i changed them to make sure they work and i didn't find anythign
out of the norm